Network Skim: 1
Removed the following categories:
fiat-accepted:  This category is redundant because although we have a Lux-specific one, this page is about the network which will contain many dApps that in-and-of-may not have an onramp.

This is better suited to its own page.

Run the Network -> Join the Network

Encouraging Lux Quest (The LRC one, not the interact-with-chain one).

Removed WAGMI:

This section is a newsletter invite for the ICO which is about to begin anyways